Move thumbnails methods
